# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: MarkFlow AI API
  description: API for the MarkFlow AI exam grading platform
  version: 1.0.0
  contact:
    name: MarkFlow AI Team

servers:
  - url: https://api.markflow.ai/v1
    description: Production
  - url: https://api-staging.markflow.ai/v1
    description: Staging
  - url: http://localhost:8080/v1
    description: Local development

security:
  - BearerAuth: []

paths:
  /health:
    get:
      summary: Health check
      operationId: healthCheck
      security: []
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  version:
                    type: string
                    example: "1.0.0"

  /exams:
    post:
      summary: Create a new exam template
      operationId: createExam
      tags: [Exams]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExamRequest"
      responses:
        "201":
          description: Exam created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    get:
      summary: List exams for the current tenant
      operationId: listExams
      tags: [Exams]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, archived]
      responses:
        "200":
          description: List of exams
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Exam"
                  total:
                    type: integer
                  page:
                    type: integer
                  pageSize:
                    type: integer

  /exams/{examId}:
    get:
      summary: Get exam by ID
      operationId: getExam
      tags: [Exams]
      parameters:
        - $ref: "#/components/parameters/ExamId"
      responses:
        "200":
          description: Exam details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
        "404":
          $ref: "#/components/responses/NotFound"

  /exams/{examId}/upload-url:
    post:
      summary: Generate a pre-signed URL for PDF upload
      operationId: generateUploadUrl
      tags: [Uploads]
      parameters:
        - $ref: "#/components/parameters/ExamId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fileName, contentType, fileSize]
              properties:
                fileName:
                  type: string
                contentType:
                  type: string
                  enum: [application/pdf]
                fileSize:
                  type: integer
                  maximum: 52428800
      responses:
        "200":
          description: Pre-signed upload URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  uploadUrl:
                    type: string
                    format: uri
                  blobPath:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time

  /exams/{examId}/upload-complete:
    post:
      summary: Notify that upload is complete, triggers ingestion pipeline
      operationId: completeUpload
      tags: [Uploads]
      parameters:
        - $ref: "#/components/parameters/ExamId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [blobPath, pageCount]
              properties:
                blobPath:
                  type: string
                pageCount:
                  type: integer
      responses:
        "202":
          description: Ingestion pipeline triggered
          content:
            application/json:
              schema:
                type: object
                properties:
                  batchId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    example: processing

  /exams/{examId}/submissions:
    get:
      summary: List submissions for an exam
      operationId: listSubmissions
      tags: [Grading]
      parameters:
        - $ref: "#/components/parameters/ExamId"
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, graded, review_needed, failed]
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Submission"
                  total:
                    type: integer

  /submissions/{submissionId}:
    get:
      summary: Get submission details
      operationId: getSubmission
      tags: [Grading]
      parameters:
        - $ref: "#/components/parameters/SubmissionId"
      responses:
        "200":
          description: Submission details with grading results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionDetail"

  /submissions/{submissionId}/grading-results:
    get:
      summary: Get grading results for a submission
      operationId: getGradingResults
      tags: [Grading]
      parameters:
        - $ref: "#/components/parameters/SubmissionId"
      responses:
        "200":
          description: Grading results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GradingResult"

  /submissions/{submissionId}/grading-results/{resultId}/override:
    patch:
      summary: Teacher overrides an AI-generated grade
      operationId: overrideGrade
      tags: [Grading]
      parameters:
        - $ref: "#/components/parameters/SubmissionId"
        - name: resultId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [overrideScore, reason]
              properties:
                overrideScore:
                  type: number
                  minimum: 0
                reason:
                  type: string
                  minLength: 1
      responses:
        "200":
          description: Grade overridden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GradingResult"

  /billing/usage:
    get:
      summary: Get current billing usage
      operationId: getUsage
      tags: [Billing]
      parameters:
        - name: month
          in: query
          schema:
            type: string
            pattern: "^\\d{4}-\\d{2}$"
      responses:
        "200":
          description: Billing usage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingUsage"

  /webhooks/stripe:
    post:
      summary: Stripe webhook endpoint
      operationId: handleStripeWebhook
      tags: [Billing]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Webhook processed

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Azure AD B2C JWT token

  parameters:
    ExamId:
      name: examId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SubmissionId:
      name: submissionId
      in: path
      required: true
      schema:
        type: string
        format: uuid

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string

    BoundingBox:
      type: object
      required: [x, y, width, height]
      properties:
        x:
          type: number
        y:
          type: number
        width:
          type: number
        height:
          type: number

    Question:
      type: object
      required: [id, type, prompt, maxMarks, boundingBox, pageNumber]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [text, equation, diagram, checkbox]
        prompt:
          type: string
        maxMarks:
          type: number
        rubric:
          type: string
        boundingBox:
          $ref: "#/components/schemas/BoundingBox"
        pageNumber:
          type: integer
        checkboxOptions:
          type: array
          items:
            type: string

    CreateExamRequest:
      type: object
      required: [title, subject, questions]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        subject:
          type: string
        questions:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Question"

    Exam:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        title:
          type: string
        subject:
          type: string
        totalMarks:
          type: number
        status:
          type: string
          enum: [draft, published, archived]
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        pageCount:
          type: integer
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Submission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        examId:
          type: string
          format: uuid
        studentId:
          type: string
        status:
          type: string
          enum: [pending, processing, graded, review_needed, failed]
        totalScore:
          type: number
        totalMaxMarks:
          type: number
        percentage:
          type: number
        averageConfidence:
          type: number
        createdAt:
          type: string
          format: date-time
        gradedAt:
          type: string
          format: date-time

    SubmissionDetail:
      allOf:
        - $ref: "#/components/schemas/Submission"
        - type: object
          properties:
            pageImages:
              type: array
              items:
                type: string
                format: uri
            gradingResults:
              type: array
              items:
                $ref: "#/components/schemas/GradingResult"

    GradingResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        submissionId:
          type: string
          format: uuid
        questionId:
          type: string
        questionType:
          type: string
          enum: [text, equation, diagram, checkbox]
        extractedAnswer:
          type: string
        aiScore:
          type: number
        maxMarks:
          type: number
        confidence:
          type: number
          minimum: 0
          maximum: 1
        feedback:
          type: string
        aiModel:
          type: string
        teacherOverride:
          type: number
          nullable: true
        overrideReason:
          type: string
          nullable: true
        gradedAt:
          type: string
          format: date-time

    BillingUsage:
      type: object
      properties:
        tenantId:
          type: string
          format: uuid
        month:
          type: string
        plan:
          type: string
        pagesScanned:
          type: integer
        aiCalls:
          type: integer
        storageBytes:
          type: integer